fix(nlu): negative levenshtein similarity + rm strings from toolkit
